requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
playwright==1.48.0
aiofiles==24.1.0
razorpay==1.4.2
cryptography==43.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.1.1
